Refactor API workflow tests and remove deprecated files
