+++ tests/workflows/basic_api_node/tests/test_workflow.py
@@ -1,12 +1,15 @@
-import requests_mock
+from unittest import mock
+
 import requests_mock.mocker
+
+from vellum.workflows.references.vellum_secret import VellumSecretReference
 
 from tests.workflows.basic_api_node.workflow import SimpleAPIWorkflow
 
 
 def test_run_workflow__happy_path(requests_mock: requests_mock.mocker.Mocker):
     # GIVEN an API request that will return a 200 OK response
-    response_mock = requests_mock.post(
+    requests_mock.post(
         "https://api.vellum.ai",
         json={"data": [1, 2, 3]},
         headers={"X-Response-Header": "bar"},
@@ -17,7 +20,9 @@ def test_run_workflow__happy_path(requests_mock: requests_mock.mocker.Mocker):
     workflow = SimpleAPIWorkflow()
 
     # WHEN we run the workflow
-    terminal_event = workflow.run()
+    with mock.patch.object(VellumSecretReference, "resolve") as mocked_resolve:
+        mocked_resolve.return_value = "SECRET_VALUE"
+        terminal_event = workflow.run()
 
     # THEN we should see the expected outputs
     assert terminal_event.name == "workflow.execution.fulfilled"
@@ -26,7 +31,3 @@ def test_run_workflow__happy_path(requests_mock: requests_mock.mocker.Mocker):
         "headers": {"X-Response-Header": "bar"},
         "status_code": 200,
     }
-
-    # AND the mock should have been called with the expected body
-    assert response_mock.last_request
-    assert response_mock.last_request.json() == {"key": "value"}
